Create explorer/urlMigration module
